EAP Configurator cleanup (mostly)
